Mejoras

+++ src/api/routes/zapatillas.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, jsonify, request
 from flask_sqlalchemy import SQLAlchemy
-from api.models import db, Zapatilla, Marca, Modelo,User
+from api.models import db, Zapatilla, Marca, Modelo,User, TallaStock
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from sqlalchemy.orm import joinedload
 
@@ -29,9 +29,10 @@ def add_zapatillas():
     if not data:
         print("No se ha enviado ningun dato")
         return jsonify({"msg": "No se ha enviado ningun dato"}), 400
-    if 'marca' not in data or 'nombre' not in data or 'talla' not in data or 'precio' not in data or 'oferta' not in data or 'genero' not in data or 'stock' not in data:
-        print("Faltan datos obligatorios")
+    required_fields = ['marca', 'nombre', 'precio', 'oferta', 'genero', 'tallas']
+    if not all(field in data for field in required_fields):
         return jsonify({"error": "Faltan datos obligatorios"}), 400
+    
     try:
         marca = Marca.query.filter_by(nombre=data['marca']).first()
         if not marca:
@@ -53,8 +54,18 @@ def add_zapatillas():
         else:
             print("El modelo ya existe")
             return jsonify({"msg": "El modelo ya existe"}), 409
-        zapatilla = Zapatilla(tallas=data['talla'],stock=data['stock'], modelo_id=modelo.id)
+        zapatilla = Zapatilla( modelo_id=modelo.id)
         db.session.add(zapatilla)
+        db.session.commit()
+
+        for entry in data['tallas']:
+            talla_stock = TallaStock(
+                zapatilla_id=zapatilla.id,
+                talla=entry['talla'],
+                stock=entry['stock']
+            )
+            db.session.add(talla_stock)
+
         db.session.commit()
         return jsonify(zapatilla.serialize()), 201
     except Exception as e:
@@ -151,12 +162,18 @@ def delete_zapatilla(zapatillas_id):
             return jsonify({"msg": "Zapatilla no encontrada"}), 404
         modelo = Modelo.query.filter_by(id=zapatilla.modelo.id).first()
         if not modelo:
-            return jsonify({"msg": "Modelo no encontrado"}), 404    
-        db.session.delete(modelo)
+            return jsonify({"msg": "Modelo no encontrado"}), 404
+        talla = TallaStock.query.filter_by(zapatilla_id=zapatillas_id).all()
+        if not talla:
+            return jsonify({"msg": "No hay tallas asociadas a esta zapatilla"}), 404
+        for talla_stock in talla:
+            db.session.delete(talla_stock)   
         db.session.delete(zapatilla)
+        db.session.delete(modelo)
         db.session.commit()
         return jsonify({"msg": "Zapatilla eliminada"}), 200
     except Exception as e:
+        print(f"Error al buscar modelo: {e}")
         db.session.rollback()
         return jsonify({"error": str(e)}), 500
 @zapatillas_bp.route('/modelo/<string:modelo_name>', methods=['GET'])
